Eliminado todo el prop drilling
