+++ src/App.tsx
@@ -1,69 +1,15 @@
 /* eslint-disable react/react-in-jsx-scope */
-import React, { useContext, useState } from 'react'
 import './App.css'
 import { Todos } from './Componentes/Todos'
-import { FilterValue, TodoContextType, type TodoId, type Todo as TodoType } from './types'
-import { TODO_FILTERS } from './const'
 import { Footer } from './Componentes/Footer'
 import { Header } from './Componentes/Header'
-import { TodoContext } from './contexts/todoContext'
 
 const App = (): React.JSX.Element => {
-  const { todos, setTodos } = useContext<TodoContextType>(TodoContext)
-  const [filterSelected, setFilterSelected] = useState<FilterValue>(TODO_FILTERS.ALL)
-
-  const handleRemove = ({ id }: TodoId): void => {
-    const newTodos = todos.filter(todo => todo.id !== id)
-    setTodos(newTodos)
-  }
-
-  const handleComplete = ({ id, completed }: Pick<TodoType, 'id' | 'completed'>): void => {
-    const newTodos = todos.map(todo => {
-      if (todo.id === id) {
-        return {
-          ...todo,
-          completed
-        }
-      }
-      return todo
-    })
-
-    setTodos(newTodos)
-  }
-
-  const handleFilterChange = (filter: FilterValue): void => {
-    setFilterSelected(filter)
-  }
-
-  const handleRemoveAllCompleted = (): void => {
-    const newTodos = todos.filter(todo => !todo.completed)
-    setTodos(newTodos)
-  }
-
-  const activeCount = todos.filter(todo => !todo.completed).length
-  const completedCount = todos.length - activeCount
-
-  const filteredTodos = todos.filter(todo => {
-    if (filterSelected === TODO_FILTERS.ACTIVE) return !todo.completed
-    if (filterSelected === TODO_FILTERS.COMPLETED) return todo.completed
-    return todo
-  })
-
   return (
     <div className='todoapp'>
       <Header />
-      <Todos
-        onToggleCompleted={handleComplete}
-        onRemoveTodo={handleRemove}
-        todos={filteredTodos}
-      />
-      <Footer
-        activeCount={activeCount}
-        completedCount={completedCount}
-        filterSelected={filterSelected}
-        onClearCompleted={handleRemoveAllCompleted}
-        handleFilterChange={handleFilterChange}
-      />
+      <Todos />
+      <Footer />
     </div>
   )
 }
