+++ src/Componentes/Todo.tsx
@@ -1,12 +1,32 @@
-import React from 'react'
+import React, { useContext } from 'react'
 import { TodoId, Todo as TodoType } from '../types'
+import { TodoContext } from '../contexts/todoContext'
 
 interface Props extends TodoType {
-  onRemoveTodo: ({ id }: TodoId) => void
-  onToggleCompleted: ({ id, completed }: Pick<TodoType, 'id' | 'completed'>) => void
 }
 
-export const Todo: React.FC <Props> = ({ id, title, completed, onRemoveTodo, onToggleCompleted }) => {
+export const Todo: React.FC <Props> = ({ id, title, completed }) => {
+  const { todos, setTodos } = useContext(TodoContext)
+
+  const handleRemove = ({ id }: TodoId): void => {
+    const newTodos = todos.filter(todo => todo.id !== id)
+    setTodos(newTodos)
+  }
+
+  const handleComplete = ({ id, completed }: Pick<TodoType, 'id' | 'completed'>): void => {
+    const newTodos = todos.map(todo => {
+      if (todo.id === id) {
+        return {
+          ...todo,
+          completed
+        }
+      }
+      return todo
+    })
+
+    setTodos(newTodos)
+  }
+
   return (
     <div className='view'>
       <input
@@ -14,15 +34,13 @@ export const Todo: React.FC <Props> = ({ id, title, completed, onRemoveTodo, onT
         checked={completed}
         type='checkbox'
         onChange={(event) => {
-          onToggleCompleted({ id, completed: event.target.checked })
+          handleComplete({ id, completed: event.target.checked })
         }}
       />
       <label>{title}</label>
       <button
         className='destroy'
-        onClick={() => {
-          onRemoveTodo({ id })
-        }}
+        onClick={() => { handleRemove({ id }) }}
       />
     </div>
