+++ src/Componentes/Todos.tsx
@@ -1,11 +1,20 @@
-import React from 'react'
-import { type Props } from '../types'
+import React, { useContext } from 'react'
 import { Todo } from './Todo'
+import { TodoContext } from '../contexts/todoContext'
+import { TODO_FILTERS } from '../const'
 
-export const Todos: React.FC<Props> = ({ todos, onRemoveTodo, onToggleCompleted }) => {
+export const Todos: React.FC = () => {
+  const { todos, filterSelected } = useContext(TodoContext)
+
+  const filteredTodos = todos.filter(todo => {
+    if (filterSelected === TODO_FILTERS.ACTIVE) return !todo.completed
+    if (filterSelected === TODO_FILTERS.COMPLETED) return todo.completed
+    return todo
+  })
+
   return (
     <ul className='todo-list'>
-      {todos.map(todo => (
+      {filteredTodos.map(todo => (
         <li
           key={todo.id}
           className={`${todo.completed ? 'completed' : ''} `}
@@ -15,8 +24,6 @@ export const Todos: React.FC<Props> = ({ todos, onRemoveTodo, onToggleCompleted 
             id={todo.id}
             title={todo.title}
             completed={todo.completed}
-            onRemoveTodo={onRemoveTodo}
-            onToggleCompleted={onToggleCompleted}
           />
         </li>
       ))}
