+++ src/contexts/todoContext.tsx
@@ -1,6 +1,7 @@
 import { createContext, useState } from 'react'
 import { mockTodos } from '../Mocks/mockTodos'
-import { Todo, TodoContextType } from '../types'
+import { FilterValue, Todo, TodoContextType } from '../types'
+import { TODO_FILTERS } from '../const'
 
 interface Props {
   children: React.ReactNode
@@ -8,16 +9,21 @@ interface Props {
 
 export const TodoContext = createContext<TodoContextType>({
   todos: [],
-  setTodos: () => {}
+  setTodos: () => {},
+  filterSelected: TODO_FILTERS.ALL,
+  setFilterSelected: () => {}
 })
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export const TodoProvider: React.FC<Props> = ({ children }) => {
   const [todos, setTodos] = useState<Todo[]>(mockTodos)
+  const [filterSelected, setFilterSelected] = useState<FilterValue>(TODO_FILTERS.ALL)
   return (
     <TodoContext.Provider value={{
       todos,
-      setTodos
+      setTodos,
+      filterSelected,
+      setFilterSelected
     }}
     >
       {children}
